+++ packages/Contracts/src/Utils/Random/Randomizer.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Aedart\Contracts\Utils\Random;
+
+use Aedart\Contracts\Utils\HasDriver;
+
+/**
+ * Randomizer
+ *
+ * @author Alin Eugen Deac <ade@rspsystems.com>
+ * @package Aedart\Contracts\Utils\Random
+ */
+interface Randomizer extends HasDriver
+{
+}