+++ packages/Utils/src/Math.php
@@ -2,6 +2,10 @@
 
 namespace Aedart\Utils;
 
+use Aedart\Contracts\Utils\Random\NumericRandomizer;
+use Aedart\Contracts\Utils\Random\Type;
+use Aedart\Utils\Random\Factory;
+use Random\Engine;
 use RuntimeException;
 use Throwable;
 
@@ -14,6 +18,20 @@ use Throwable;
 class Math
 {
     /**
+     * Returns a new Numeric Randomizer instance
+     *
+     * @param Engine|null $engine [optional]
+     *
+     * @return NumericRandomizer
+     */
+    public static function randomizer(Engine|null $engine = null): NumericRandomizer
+    {
+        return Factory::make(Type::Numeric, $engine);
+    }
+
+    /**
+     * @deprecated Since v8.0 - Use \Aedart\Utils\Math::randomizer()->int() instead.
+     *
      * Returns a random cryptographically secure, uniformly selected number
      *
      * @see https://www.php.net/manual/en/function.random-int
@@ -50,7 +68,7 @@ class Math
     {
         // Source is from php.net's documentation:
         // @see https://www.php.net/manual/en/function.mt-srand.php#refsect1-function.mt-srand-examples
-        list($usec, $sec) = explode(' ', microtime());
+        [$usec, $sec] = explode(' ', microtime());
         return (int) $sec + $usec * 1000000;
     }