+++ packages/Utils/src/Str.php
@@ -2,8 +2,12 @@
 
 namespace Aedart\Utils;
 
+use Aedart\Contracts\Utils\Random\StringRandomizer;
+use Aedart\Contracts\Utils\Random\Type;
+use Aedart\Utils\Random\Factory;
 use Illuminate\Support\Str as BaseStr;
 use Illuminate\Support\Stringable;
+use Random\Engine;
 
 /**
  * String utility
@@ -13,6 +17,18 @@ use Illuminate\Support\Stringable;
  */
 class Str extends BaseStr
 {
+    /**
+     * Returns a new String Randomizer instance
+     *
+     * @param Engine|null $engine [optional]
+     *
+     * @return StringRandomizer
+     */
+    public static function randomizer(Engine|null $engine = null): StringRandomizer
+    {
+        return Factory::make(Type::String, $engine);
+    }
+
     /**
      * Converts slug back to words
      *