+++ tests/Unit/Utils/StrTest.php
@@ -3,6 +3,7 @@
 
 namespace Aedart\Tests\Unit\Utils;
 
+use Aedart\Contracts\Utils\Random\StringRandomizer;
 use Aedart\Testing\Helpers\ConsoleDebugger;
 use Aedart\Testing\TestCases\UnitTestCase;
 use Aedart\Utils\Str;
@@ -19,6 +20,18 @@ use Aedart\Utils\Str;
  */
 class StrTest extends UnitTestCase
 {
+    /**
+     * @test
+     *
+     * @return void
+     */
+    public function returnsRandomizer(): void
+    {
+        $randomizer = Str::randomizer();
+
+        $this->assertInstanceOf(StringRandomizer::class, $randomizer);
+    }
+
     /**
      * @test
      */